+++ src/Laravel/AppPool.php
@@ -3,6 +3,7 @@
 namespace HuangYi\Shadowfax\Laravel;
 
 use HuangYi\Shadowfax\Contracts\AppPool as PoolContract;
+use HuangYi\Shadowfax\Events\AppPoppedEvent;
 use HuangYi\Shadowfax\Events\AppPushingEvent;
 use HuangYi\Shadowfax\HasEventDispatcher;
 use Illuminate\Contracts\Container\Container;
@@ -49,6 +50,8 @@ class AppPool implements PoolContract
     public function pop(): Container
     {
         Facade::clearResolvedInstances();
+
+        $this->dispatch(AppPoppedEvent::class, $this->app);
 
         return $this->app;
     }